moved part of server.js code to index.js
